# api/docs/swagger.yaml
definitions:
...
  item.AddItemCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  item.AddItemCategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  item.AddItemRequest:
    properties:
      category_id:
        type: string
      condition:
        type: string
      description:
        type: string
      name:
        type: string
      swap_preference:
        type: string
      user_id:
        type: string
    type: object
...
  item.CreateEcoChallengeRequest:
    properties:
      description:
        type: string
      end_date:
        type: string
      reward_points:
        type: integer
      start_date:
        type: string
      title:
        type: string
    type: object
  item.CreateEcoChallengeResponse:
    properties:
      challenge:
        $ref: '#/definitions/item.EcoChallenge'
    type: object
  item.CreateEcoTipRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  item.CreateEcoTipResponse:
    properties:
      tip:
        $ref: '#/definitions/item.EcoTip'
    type: object
  item.DeleteItemResponse:
    properties:
      message:
        type: string
    type: object
  item.EcoChallenge:
    properties:
      created_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: string
      reward_points:
        type: integer
      start_date:
        type: string
      title:
        type: string
    type: object
  item.EcoTip:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  item.GetEcoTipsRequest:
    properties:
      limit:
        type: integer
      page:
        type: integer
    type: object
  item.GetEcoTipsResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      tips:
        items:
          $ref: '#/definitions/item.EcoTip'
        type: array
      total:
        type: integer
    type: object
...
  item.ItemResponse:
    properties:
      category_id:
        type: string
      condition:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: string
      swap_preference:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  item.ListItemsRequest:
    properties:
      limit:
        type: integer
      page:
        type: integer
    type: object
  item.ListItemsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/item.ItemResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  item.ParticipateEcoChallengeRequest:
    properties:
      challenge_id:
        type: string
      user_id:
        type: string
    type: object
  item.ParticipateEcoChallengeResponse:
    properties:
      challenge_id:
        type: string
      joined_at:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
...
  item.SearchItemsRequest:
    properties:
      category:
        type: string
      condition:
        type: string
      limit:
        type: integer
      page:
        type: integer
    type: object
...
  item.UpdateEcoChallengeProgressRequest:
    properties:
      challenge_id:
        type: string
      recycled_items_count:
        type: integer
    type: object
  item.UpdateEcoChallengeProgressResponse:
    properties:
      challenge_id:
        type: string
      recycled_items_count:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  item.UpdateItemRequest:
    properties:
      category_id:
        type: string
      condition:
        type: string
      description:
        type: string
      item_id:
        type: string
      name:
        type: string
      swap_preference:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Gateway of User Item System
  title: User Item System
  version: "1.0"
paths:
  /item-system/category/catogories:
    post:
      description: Inserts new item category info into item_categories table in PostgreSQL
      parameters:
      - description: New data
        in: body
        name: new_data
        required: true
        schema:
          $ref: '#/definitions/item.AddItemCategoryRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.AddItemCategoryResponse'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while adding item category
          schema:
            type: string
      summary: Adds an item category
      tags:
      - item
  /item-system/eco-tips:
    get:
      description: Retrieves all eco tips info from PostgreSQL
      parameters:
      - description: Request data
        in: body
        name: new_data
        required: true
        schema:
          $ref: '#/definitions/item.GetEcoTipsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.GetEcoTipsResponse'
        "500":
          description: Server error while getting eco tips
          schema:
            type: string
      summary: Gets all eco tips
      tags:
      - eco_tip
    post:
      description: Inserts new eco tip info into eco_tips table in PostgreSQL
      parameters:
      - description: New data
        in: body
        name: new_data
        required: true
        schema:
          $ref: '#/definitions/item.CreateEcoTipRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.CreateEcoTipResponse'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while creating eco tip
          schema:
            type: string
      summary: Creates a new eco tip
      tags:
      - eco_tip
  /item-system/ecosystem/eco-challenge:
    post:
      description: Inserts new eco challenge info into eco_challenges table in PostgreSQL
      parameters:
      - description: New data
        in: body
        name: new_data
        required: true
        schema:
          $ref: '#/definitions/item.CreateEcoChallengeRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.CreateEcoChallengeResponse'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while creating eco challenge
          schema:
            type: string
      summary: Creates a new eco challenge
      tags:
      - eco_challenge
  /item-system/ecosystem/participate:
    post:
      description: Inserts new participation info into challenge_participations table
        in PostgreSQL
      parameters:
      - description: New data
        in: body
        name: new_data
        required: true
        schema:
          $ref: '#/definitions/item.ParticipateEcoChallengeRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.ParticipateEcoChallengeResponse'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while participating in eco challenge
          schema:
            type: string
      summary: Participates in an eco challenge
      tags:
      - eco_challenge
  /item-system/ecosystem/update:
    put:
      description: Updates progress info in challenge_participations table in PostgreSQL
      parameters:
      - description: New data
        in: body
        name: new_data
        required: true
        schema:
          $ref: '#/definitions/item.UpdateEcoChallengeProgressRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.UpdateEcoChallengeProgressResponse'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while updating eco challenge progress
          schema:
            type: string
      summary: Updates progress in an eco challenge
      tags:
      - eco_challenge
  /item-system/items:
    post:
      description: Retrieves all items info from items table in PostgreSQL
      parameters:
      - description: list item data
        in: body
        name: update_data
        required: true
        schema:
          $ref: '#/definitions/item.ListItemsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.ListItemsResponse'
        "500":
          description: Server error while listing items
          schema:
            type: string
      summary: Lists all items
      tags:
      - item
  /item-system/items/{item_id}:
    delete:
      description: Deletes item from items table in PostgreSQL
      parameters:
      - description: Item ID
        in: path
        name: item_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.DeleteItemResponse'
        "400":
          description: Invalid item ID
          schema:
            type: string
        "500":
          description: Server error while deleting item
          schema:
            type: string
      summary: Deletes an item
      tags:
      - item
    get:
      description: Retrieves item info from items table in PostgreSQL
      parameters:
      - description: Item ID
        in: path
        name: item_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.ItemResponse'
        "400":
          description: Invalid item ID
          schema:
            type: string
        "500":
          description: Server error while getting item
          schema:
            type: string
      summary: Gets an item
      tags:
      - item
    put:
      description: Updates item info in items table in PostgreSQL
      parameters:
      - description: Item ID
        in: path
        name: item_id
        required: true
        type: string
      - description: Updated item data
        in: body
        name: update_data
        required: true
        schema:
          $ref: '#/definitions/item.UpdateItemRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.ItemResponse'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while updating item
          schema:
            type: string
      summary: Updates an item
      tags:
      - item
  /item-system/items/addItem:
    post:
      description: Inserts new item info into items table in PostgreSQL
      parameters:
      - description: New item data
        in: body
        name: new_data
        required: true
        schema:
          $ref: '#/definitions/item.AddItemRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.ItemResponse'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while adding item
          schema:
            type: string
      summary: Adds a new item
      tags:
      - item
  /item-system/items/search:
    post:
      description: Searches items info in items table in PostgreSQL
      parameters:
      - description: list item data
        in: body
        name: update_data
        required: true
        schema:
          $ref: '#/definitions/item.SearchItemsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.ListItemsResponse'
        "500":
          description: Server error while searching items
          schema:
            type: string
      summary: Searches for items
      tags:
      - item
...
swagger: "2.0"